Moving metadata operations completely to provisioner Related refactoring
