+++ spring-cloud-stream-binder-kafka-0.9-test/src/test/java/org/springframework/cloud/stream/binder/kafka/Kafka_09_BinderTests.java
@@ -40,7 +40,6 @@ import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
 import org.springframework.kafka.support.KafkaHeaders;
 import org.springframework.kafka.test.core.BrokerAddress;
 import org.springframework.kafka.test.rule.KafkaEmbedded;
-import org.springframework.retry.RetryOperations;
 
 /**
  * Integration tests for the {@link KafkaMessageChannelBinder}.
@@ -91,11 +90,6 @@ public class Kafka_09_BinderTests extends KafkaBinderTests {
 	@Override
 	protected int partitionSize(String topic) {
 		return consumerFactory().createConsumer().partitionsFor(topic).size();
-	}
-
-	@Override
-	protected void setMetadataRetryOperations(Binder binder, RetryOperations retryOperations) {
-		((Kafka09TestBinder) binder).getBinder().setMetadataRetryOperations(retryOperations);
 	}
 
 	@Override
